Restructure pages on icons
We previously had one page on icon usage and another on icon
creation. However, this got a bit unclear with regards to what
types of icon to use when.

This change restructures the pages so we have one on app icons
and one on symbolics. This makes the types of icons we have
simpler, and reinorces the idea that symbolics generally get
reused, whereas every app icon is unique.

As part of this, also fix #51.