Add content from the accessibility UI guidelines
The old accessibility developers guide has a section on UI
guidelines. Really, the HIG ought to cover accessibility, so
that accessibility is built in by default, and so that designers
and developers don't need to use a separate guide.

This commit adds a bunch of material from that guide. There might
be some more to add, but I think we cover most points now.