build-system: add Makefile and allow parallel compilation
At the moment you must go to the commandline, type in ./00localbuild.sh,
wait much seconds, even when you have only changed one file, and then open
build/index.html with your browser.

With gnome-builder it would be much easier to simply click the "run"
or "build" button, and it would compile and even start the browser.
Also only files that have been changed are getting recompiled.

Another change is that the old buildscript builds now in parallel,
so its much faster. This script is left there, because maybe someone
prefer it and the two solutions can exist together.