placeholders: simplify
In the past we had two types of placeholder, which could be used
in the same view depending on the cirmustance. This seemed a bit
over-complex.

Sticking with the two styles of placeholder, but using them for
different types of view is simpler and has much of the same
effect.