New pages on navigation and browsing
We had a design pattern called stacks, but it was too easily
confused with GtkStack. It combined two things: a type of structure
with a means of navigating it. It's actually helpful to split these
up and deal with them separately, hence the two new pages, one on
navigation structures and one on browsing.

Fixes #84