Recommend grid views over flow boxes
As discussed in #85, grid views are a better option for content
grids. This commit drops the page on flow boxes for now and
recommends grid views instead. We may want to reference flow boxes
in a future "tiles" design pattern.

Other changes included here: refer to "lists" as "boxed lists", in
order to better differentiate from list views.

Fixes #85