reference/palette: Improve table appearance
Swatches are now circular, table spans the full width of the page,
removed extraneous space between the bottom of swatches & the
bottom of the table row, and the swatches are now mobile friendly.

CSS `aspect-ratio` is used by default to handle swatch sizing; if
this isn't supported, it falls back to statically defined
width/height.